+++ elements/pfe-accordion/demo/pfe-accordion.js
@@ -1,3 +1,4 @@
+import '@patternfly/pfe-accordion';
 import '@patternfly/pfe-band';
 import '@patternfly/pfe-button';
 import '@patternfly/pfe-card';