+++ elements/pfe-health-index/CHANGELOG.md
@@ -1,20 +1,30 @@
 # @patternfly/pfe-health-index
 
+## 2.0.0-next.1
+
+### Patch Changes
+
+- 447b2d75: Remove `esbuild` export condition, as this anyways was a runtime error
+- Updated dependencies [447b2d75]
+  - @patternfly/pfe-core@2.0.0-next.3
+
 ## 2.0.0-next.0
+
 ### Major Changes
 
 - dd88587e: ## 🔥 Migrate to Lit
-  
+
   This release migrates `<pfe-health-index>` to LitElement.
-  
+
   ### NEW: CSS Shadow Parts!
+
   - Adds `container` and `box` CSS parts to `<pfe-health-index>`
-  
+
   ### Breaking Changes
+
   - Initial render is now [asynchronous](https://lit.dev/docs/components/lifecycle/#reactive-update-cycle).
     If your code assumes that shadow DOM is ready once the element is constructed, update it to `await element.updateComplete`
-  
-  
+
   See [docs](https://patternflyelements.org/components/health-index/) for more info
 
 ### Patch Changes