+++ elements/pfe-readtime/CHANGELOG.md
@@ -1,17 +1,26 @@
 # @patternfly/pfe-readtime
 
+## 2.0.0-next.1
+
+### Patch Changes
+
+- 447b2d75: Remove `esbuild` export condition, as this anyways was a runtime error
+- Updated dependencies [447b2d75]
+  - @patternfly/pfe-core@2.0.0-next.3
+
 ## 2.0.0-next.0
+
 ### Major Changes
 
 - d7a45163: ## 🔥 Migrate to Lit
-  
+
   This release migrates `<pfe-readtime>` to LitElement.
-  
+
   ### Breaking Changes
+
   - Initial render is now [asynchronous](https://lit.dev/docs/components/lifecycle/#reactive-update-cycle).
     If your code assumes that shadow DOM is ready once the element is constructed, update it to `await element.updateComplete`
-  
-  
+
   See [docs](https://patternflyelements.org/components/readtime/) for more info
 
 ### Patch Changes